Moved memory check to System
