+++ src/sys/system.h
@@ -58,6 +58,17 @@ class System
         bool       skip_clocks;
     };
 
+    /** A simple way to represent where the program is executing from
+     * 
+     */
+    enum ProgramMemory 
+    {
+      INTERNAL_FLASH = 0,
+      AXI_SRAM,
+      QSPI,
+      INVALID_ADDRESS,
+    };
+
     System() {}
     ~System() {}
 
@@ -150,10 +161,24 @@ class System
      */
     const Config& GetConfig() const { return cfg_; }
 
+    /** Returns an enum representing the current (primary) memory space used
+     *  for executing the program.
+     */
+    static ProgramMemory GetProgramMemory();
+
   private:
     void   ConfigureClocks();
     void   ConfigureMpu();
     Config cfg_;
+
+    // TODO -- unify this with bootloader so we don't have reduntant values
+    static constexpr uint32_t sram_start_ = 0x24000000U;
+    static constexpr uint32_t sram_end_ = sram_start_ + 0x80000U;
+    static constexpr uint32_t qspi_start_ = 0x90040000U;
+    // TODO -- this is a bit too large:
+    static constexpr uint32_t qspi_end_ = qspi_start_ + 0x800000U;
+    static constexpr uint32_t internal_start_ = 0x08000000U;
+    static constexpr uint32_t internal_end_ = internal_start_ + 0x20000U;
 
     /** One TimerHandle to rule them all
      ** Maybe this whole class should be static.. */
